Add tests for the INTEL_performance_query extension.
Petri's piglit tests for INTEL_performance_query from last year fell
through the cracks so here's a rebased patch...

--- >8 --- (git am --scissors to remove)

As with AMD_performance_monitor, one of the challenging aspects of
testing this extension is that it defines an implementation-specific
set of groups and counters. Many of the tests here arbitrarily
operate on the counters in the first query, while a few sanity check
all counters in all queries.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>