In the end, it only made sense on GetVertexAttrib.
This is failing right now, but fwiw, raising INVALID_OPERATION
when index is zero is checked on the mesa code.

But it seems that "index 0" status depends on the opengl version.
More info on this commit:
49c71050de1e0df451e29621558b4adfefb7e098

It stopped to work. Quoting the commit message:
"In OpenGL 3.1 attribute 0 becomes non-magic, just like in
 OpenGL ES 2.0. Earlier versions of OpenGL used attribute 0
 exclusively for vertex position."

Further research is needed.