arb_timer_query: Add wall clock sanity checks
The tests so far verified that timestamp queries via glGet or
glQueryCounter were consistent but didn't check that timestamp deltas
have nanosecond units.

This tests synchronously reading gpu timestamps either side of a
cpu sleep and expects the gpu timestamp delta to be >= the cpu
sleep duration. Both glGet and glQueryCounter are tested.

Signed-off-by: Robert Bragg <robert@sixbynine.org>